Fix : member로 변경 및 point long으로 변경

+++ backend/member/src/main/java/com/paymong/member/paypoint/dto/AddPaypointResDto.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AddPaypointResDto {
-    Integer point;
+    Long point;
     String mapCode;
 }
